add -E option to filepath command (#42)
* add -E option to filepath command

-E	
Specifies the <exit-code> if HDBSQL returns an error. Values are: OKAY = 0, ERROR = 1, USAGE = 2, SQLERROR = 3, and VAR = 4.

if -E is not set when using-I option the sql will not fail on first error and still continue running the sql commands inside the file. If the last command is successfully, the whole run is market as successfully

* Create 0044-hdbsql_sqlfile_faile_on_first_error.yaml

* Update 0044-hdbsql_sqlfile_faile_on_first_error.yaml

* Update and rename 0044-hdbsql_sqlfile_faile_on_first_error.yaml to 0044-hdbsql_sqlfile_fail_on_first_error.yaml

---------

Co-authored-by: Rainer Leber <39616583+rainerleber@users.noreply.github.com>